@@ -1,27 +1,20 @@
 {
-    "homepage": "http://ubaimutl.github.io/react-portfolio",
     "name": "react-portfolio",
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-        "@steveeeie/react-page-transition": "^1.2.1",
-        "@testing-library/jest-dom": "^5.11.4",
-        "@testing-library/react": "^11.1.0",
-        "@testing-library/user-event": "^12.1.10",
-        "bootstrap": "^4.6.0",
-        "emailjs-com": "^2.6.4",
-        "gh-pages": "^3.2.3",
-        "react": "^17.0.1",
-        "react-animated-cursor": "^2.1.8",
-        "react-bootstrap": "^1.4.3",
-        "react-dom": "^17.0.1",
+        "bootstrap": "^5.2.3",
+        "emailjs-com": "^3.2.0",
+        "react": "^18.2.0",
+        "react-bootstrap": "^2.7.0",
+        "react-dom": "^18.2.0",
         "react-helmet-async": "^1.0.7",
         "react-icons": "^4.1.0",
-        "react-router-dom": "^5.2.0",
-        "react-scripts": "4.0.1",
+        "react-router-dom": "^6.6.2",
+        "react-scripts": "^5.0.1",
         "react-transition-group": "^4.4.2",
         "typewriter-effect": "^2.18.2",
-        "web-vitals": "^0.2.4"
+        "web-vitals": "^3.1.1"
     },
     "scripts": {
         "start": "react-scripts start",
