@@ -3,11 +3,11 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import {
   BrowserRouter as Router,
   useLocation,
-  withRouter,
 } from "react-router-dom";
+import withRouter from "../hooks/withRouter";
 import AppRoutes from "./routes";
 import Headermain from "../header";
-import AnimatedCursor from "react-animated-cursor";
+import AnimatedCursor  from "../hooks/AnimatedCursor";
 import "./App.css";
 
 function _ScrollToTop(props) {
